+++ examples/experimental/eval/config/scene_manipulation_config.yaml
