+++ examples/experimental/eval/scene_manipulation.py
